Initial re-structuring done
