+++ TensorFlow/README.md
@@ -1,4 +1,30 @@
-# TensorFlow
+# Folder Structure
 
-Un-official TensorFlow components, like the Object Detection API, can be downloaded into this folder for
-installation purposes. All the contents of this folder will be ignored by `.gitignore`.
+* **[object_detection_api](./object_detection_api/README.md)** a directory where you can install the TensorFlow Object Detection API
+* **[large_files](./large_files/README.md)** some of the examples download training and testing images to this folder
+
+---
+
+# TensorFlow 2 Basics
+
+* [Data type basics](./Basics_data_type_tutorial.ipynb)
+* [Automatic differentiation](./Basics_automatic_differentiation.ipynb)
+* [TensorFlow 2 functional API](./Basics_functional_api.ipynb)
+
+---
+
+# Deep Learning Examples
+
+* [Simple classification example](./simple_classification_example.ipynb)
+* [Simple linear regression example](./simple_linear_regression_example.ipynb)
+* [ANN handwritten number classification example](./ANN-handwritten-digits-classification.ipynb)
+* [CNN cifar 10 classification example](./CNN_cifar_10_classification.ipynb)
+* [VGG transfer learning example](./VGG_transfer_learning_fruit_data.ipynb)
+* [ResNet transfer learning example](./ResNet_transfer_learning_fruit_data.ipynb)
+* [ResNet implementation in TensorFlow/Keras](./ResNet_implementation.ipynb)
+* [Image embedding example using Fiftyone](./Fiftyone_image_embedding_example.ipynb)
+* [Tensorflow Object Detection API (SSD) example](./SSD_tensorflow2_object_detection_example.ipynb)
+* [Neural style transfer part 1, reproducing content](./Neural_style_transfer_part_1.ipynb)
+* [Neural style transfer part 2, reproducing style](./Neural_style_transfer_part_2.ipynb)
+* [Neural style transfer part 3, reproducing content and style](./Neural_style_transfer_part_3.ipynb)
+* [Class activation maps](./class_activation_maps.ipynb)
