+++ TensorFlow/large_files/README.md
@@ -1,0 +1,4 @@
+# Large Files
+
+Test images and videos will be downloaded to this folder by some of the examples for testing purposes.
+All the contents of this folder will be ignored by `.gitignore`.
