+++ conda/tensorflow2_gpu.yml
@@ -4,15 +4,18 @@ channels:
     - conda-forge
     - anaconda
 dependencies:
-    - python=3.8
-    - jupyter
-    - scikit-learn
-    - matplotlib
+    - python=3.9
+    - cudatoolkit=11.2
+    - cudnn=8.1.0
     - pip
     - pip:
-        - tensorflow-gpu==2.8.0
+        - tensorflow-gpu==2.10.0
         - tensorflow-datasets
         - opencv-python
         - gdown
         - fiftyone
+        - jupyter
+        - scikit-learn
+        - matplotlib
+
