--- large_files/.gitignore
@@ -1,12 +1,0 @@
-# Ignore everything
-*
-
-# Apart from .gitignore and README.md
-!.gitignore
-!README.md
-
-# And the following images
-!van_gogh.jpeg
-!large_files/picasso.jpg
-!picasso-weeping-woman.jpg
-
